requirements.txt
# REACH Code Visualizer - Dependencies
# Phase 1: Core Parsing

# Web server (for later phases)
flask==3.0.0
flask-socketio==5.3.0

# File watching
watchdog==3.0.0

# Graph operations
networkx==3.2.1

# Configuration
pyyaml==6.0.1

# Environment variables
python-dotenv==1.0.0

# CLI utilities
click==8.1.7
rich==13.7.0

# Development/Testing
pytest==7.4.4
